UPDATE: Responsive Footer bar.
### Added:

- Without changes.

### Changed

- Footer bar adjusted for mobile devices and reorganized for a better presentation on desktop view.

### Deprecated

- Without changes.

### Removed

- Without changes.

### Fixed

- Without changes.

### Security

- Without changes.
