+++ src/components/layout/header/header.css
@@ -11,7 +11,14 @@
     width: 100%;
     max-height: 50px;
     max-width: 290px;
+    filter: drop-shadow(10px 10px 10px rgba(0, 0, 0, 0.4));
+    transition: transform 0.5s ease-in-out;
 }
+
+.layout_header_LogoRDB:hover {
+    /*transform: perspective(500px) translateZ(0) rotateY(-20deg);*/
+}
+
 
 .header_right {
     float: left;
